# docs/swagger.yaml
definitions:
  github_com_maczh_mgin_models.Result-any:
    properties:
      data: {}
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
...
  github_com_maczh_mgin_models_sys.SysUserExt:
    properties:
      departmentId:
        type: integer
      id:
        type: integer
      positionId:
        type: integer
      roleId:
        type: integer
      userId:
        type: integer
    type: object
...
  github_com_maczh_mgin_models_sys_request.CreateSysUserExtReq:
    properties:
      departmentId:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      positionId:
        type: integer
      roleId:
        type: integer
      userId:
        type: integer
    required:
    - userId
    type: object
...
  models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/v1/user/ext/add:
    post:
      consumes:
      - application/json
      description: 创建用户的扩展信息，如部门、角色、职位等
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建用户扩展信息请求参数
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.CreateSysUserExtReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回创建的用户扩展信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或创建用户扩展信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 创建用户扩展信息
      tags:
      - 用户扩展信息
  /api/v1/user/ext/del:
    post:
      consumes:
      - application/json
      description: 根据用户 ID 删除单个用户的扩展信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或删除用户扩展信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除单个用户扩展信息
      tags:
      - 用户扩展信息
  /api/v1/user/ext/get:
    get:
      consumes:
      - application/json
      description: 根据用户 ID 获取单个用户的扩展信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户扩展信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取用户扩展信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取单个用户扩展信息
      tags:
      - 用户扩展信息
  /api/v1/user/ext/list:
    get:
      consumes:
      - application/json
      description: 根据查询条件获取用户扩展信息的列表，支持分页
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: departmentId
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: positionId
        type: integer
      - in: query
        name: roleId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户扩展信息列表和总数
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取用户扩展信息列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取用户扩展信息列表
      tags:
      - 用户扩展信息
  /api/v1/user/ext/update:
    post:
      consumes:
      - application/json
      description: 更新用户已有的扩展信息，如部门、角色、职位等
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新用户扩展信息请求参数
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysUserExt'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败、获取用户扩展信息失败或更新用户扩展信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 更新用户扩展信息
      tags:
      - 用户扩展信息
...
swagger: "2.0"